adding if is/not installed

+++ pkg-install.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
+# Author: Edvin Dunaway
+# Contact: edvin@eddinn.net
+# Version: 0.1.5
+
+# TODO:
+# Server/Workstation install
+# Combine with setup.sh
+# Add gnome-shell-extensions
+
+# Upgrade the system
 sudo apt dist-upgrade
 
+# Define user Apt packages to install
 apt_packages=(
  audacity
  bash-completion
@@ -38,7 +49,6 @@ apt_packages=(
  snapd
  steam
  stow
- teamviewer
  unattended-upgrades
  unzip
  vim
@@ -47,6 +57,7 @@ apt_packages=(
  zsh
 )
 
+# Define Python3 Pip packages to install
 pip_packages=(
  setuptools
  wheel
@@ -57,9 +68,9 @@ pip_packages=(
  testresources
 )
 
+# Define Snap packages to install
 snap_packages=(
  code
- discord
  gitkraken
  spotify
 )
@@ -71,12 +82,24 @@ echo "Upgrading pip3 and installing Python3 packages"
 sudo -H pip3 install pip --upgrade
 sudo -H pip3 install "${pip_packages[@]}"
 
-echo "Installing google Chrome"
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo dpkg -i google-chrome-stable_current_amd64.deb
-rm -Rf google-chrome-stable_current_amd64.deb
+echo "Installing Google Chrome"
+if [ "$(dpkg-query -W -f='${Status}' google-chrome-stable 2>/dev/null | grep -c "ok installed")" -eq 0 ];
+then
+  wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+  sudo dpkg -i google-chrome-stable_current_amd64.deb
+  rm -Rf google-chrome-stable_current_amd64.deb
+fi
+
+echo "Installing TeamViewer"
+if [ "$(dpkg-query -W -f='${Status}' teamviewer 2>/dev/null | grep -c "ok installed")" -eq 0 ];
+then
+  wget https://download.teamviewer.com/download/linux/teamviewer_amd64.deb
+  sudo dpkg -i teamviewer_amd64.deb
+  rm -Rf teamviewer_amd64.deb
+fi
 
 echo "Installing Snap packages"
+sudo snap install discord --classic
 for i in "${snap_packages[@]}"; do sudo snap install "$i"; done
 
 echo "All done!"
